+++ src/main/java/com/project/stms/controller/UserController.java
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.project.stms.command.UserVO;
-import com.project.stms.user.service.UserService;
+import com.project.stms.service.user.UserService;
 
 @Controller
 public class UserController {
