+++ src/main/java/com/project/stms/service/user/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.project.stms.user.service;
+package com.project.stms.service.user;
 
 import java.util.List;
